moves ApiController to base level and out of legacy folder
