+++ app/Http/Controllers/Legacy/Two/EventController.php
@@ -4,6 +4,7 @@ namespace Rogue\Http\Controllers\Legacy\Two;
 
 use Rogue\Models\Event;
 use Illuminate\Http\Request;
+use Rogue\Http\Controllers\ApiController;
 use Rogue\Http\Transformers\Legacy\Two\EventTransformer;
 
 class EventController extends ApiController
