+++ app/Http/Controllers/Legacy/Two/ReportbackController.php
@@ -5,6 +5,7 @@ namespace Rogue\Http\Controllers\Legacy\Two;
 use Rogue\Models\Post;
 use League\Fractal\Manager;
 use Illuminate\Http\Request;
+use Rogue\Http\Controllers\ApiController;
 use Rogue\Http\Transformers\Legacy\Two\ReportbackTransformer;
 use Rogue\Http\Transformers\Legacy\Two\PaginatorForAshesGallery;
