+++ app/Http/Controllers/ReactionController.php
@@ -5,8 +5,8 @@ namespace Rogue\Http\Controllers;
 use Rogue\Models\Post;
 use Rogue\Models\Reaction;
 use Illuminate\Http\Request;
+use Rogue\Http\Controllers\ApiController;
 use Rogue\Http\Transformers\ReactionTransformer;
-use Rogue\Http\Controllers\Legacy\Two\ApiController;
 
 class ReactionController extends ApiController
 {
