+++ app/Http/Controllers/SignupsController.php
@@ -5,9 +5,9 @@ namespace Rogue\Http\Controllers;
 use Rogue\Models\Signup;
 use Illuminate\Http\Request;
 use Rogue\Managers\SignupManager;
+use Rogue\Http\Controllers\ApiController;
 use Rogue\Http\Transformers\SignupTransformer;
 use Illuminate\Auth\Access\AuthorizationException;
-use Rogue\Http\Controllers\Legacy\Two\ApiController;
 use Rogue\Http\Controllers\Traits\TransformsRequests;
 
 class SignupsController extends ApiController
